Revert "fix: 불러온 회고가 아닌, 저장된 전체 회고에서 개수를 카운팅하도록 변경"
This reverts commit 2f1ae54ef20b96c4de4aa35990794587200f8c0c.
