+++ RetsTalk/RetsTalk/Retrospect/Model/SortedRetrospects.swift
@@ -16,4 +16,24 @@ struct SortedRetrospects {
     
     subscript(row: Int) -> [Retrospect] { retrospects[row] }
     subscript(row: Int, column: Int) -> Retrospect { retrospects[row][column] }
+    
+    var totalCount: Int { retrospects.flatMap { $0 }.count }
+    var datesCount: Int {
+        let allRetrospects = retrospects.flatMap { $0 }
+        let uniqueDates = Set(allRetrospects.map { retrospect in
+            let formatter = DateFormatter()
+            formatter.dateFormat = Texts.dateFormat
+            return formatter.string(from: retrospect.createdAt)
+        })
+        
+        return uniqueDates.count
+    }
+}
+
+// MARK: - Constants
+
+private extension SortedRetrospects {
+    enum Texts {
+        static let dateFormat = "yyyy-MM-dd"
+    }
 }
